Fix cdc-wdm modem reset resulting in invalid state
Clearing the MODEMMANAGER_CDCWDM_CACHE at mm_report_events_from_cache
will cause the cache to be removed at start of the ModemManager.

This will cause the hotplug event 25-modemmanager-net to be unable to
report the cdc-wdm event removed. Resulting the modem itself not to be
removed from ModemManager.

When the interfaces are added, they get linked to the old modem 'object'
resulting in a invalid state. Making it impossible to modify the object.